requirements.txt
oauthlib
requests_oauthlib
pyyaml